+++ src/getInputs.ts
@@ -1,5 +1,6 @@
 import { getInput } from '@actions/core/lib/core';
 import {
+  IssuesAddAssigneesParams,
   IssuesAddLabelsParams,
   PullsCreateParams,
   PullsCreateReviewRequestParams,
@@ -8,10 +9,13 @@ import {
 type Inputs =
   & PullsCreateParams
   & Required<
-    Omit<PullsCreateReviewRequestParams, 'pull_number' | 'team_reviewers'>
+    Omit<PullsCreateReviewRequestParams, 'pull_number'>
   >
   & Required<
     Omit<IssuesAddLabelsParams, 'issue_number'>
+  >
+  & Required<
+    Omit<IssuesAddAssigneesParams, 'issue_number'>
   >;
 
 export function getInputs(): Inputs {
@@ -20,7 +24,9 @@ export function getInputs(): Inputs {
   const base = getInput('base') || 'master';
   const draft = getInput('draft') ? JSON.parse(getInput('draft')) : undefined;
   const body = getInput('body') || undefined;
+  const assignees = getInput('assignees');
   const reviewers = getInput('reviewers');
+  const team_reviewers = getInput('team_reviewers');
   const labels = getInput('labels');
   const repository = getInput('repository');
 
@@ -40,8 +46,14 @@ export function getInputs(): Inputs {
     body,
     owner,
     repo,
+    assignees: assignees
+      ? assignees.split(',').map(assignee => assignee.trim())
+      : [],
     reviewers: reviewers
       ? reviewers.split(',').map(reviewer => reviewer.trim())
+      : [],
+    team_reviewers: team_reviewers
+      ? team_reviewers.split(',').map(reviewer => reviewer.trim())
       : [],
     labels: labels
       ? labels.split(',').map(label => label.trim())