+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Support for adding team reviewers by setting `team_reviewers`
+- Support for adding assignees by setting `assignees`
+
 ## [1.2.0] - 2021-07-26
 
 ### Added